#278 #139 Refactoring of ConfigurationRequirement Test.
