+++ algorithm_integration/src/test/java/de/metanome/algorithm_integration/configuration/ConfigurationRequirementBooleanTest.java
@@ -23,8 +23,6 @@ import de.metanome.test_helper.GwtSerializationTester;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 
@@ -32,85 +30,6 @@ import static org.mockito.Mockito.mock;
  * Tests for {@link ConfigurationRequirementBoolean}
  */
 public class ConfigurationRequirementBooleanTest {
-
-  /**
-   * Test method for {@link ConfigurationRequirementBoolean#ConfigurationRequirementBoolean(String)}
-   * <p/> The identifier should be set in the constructor and be retrievable through getIdentifier.
-   * The numberOfValues should be set to 1.
-   */
-  @Test
-  public void testConstructorGetOne() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedNumberOfValues = 1;
-
-    // Execute functionality
-    ConfigurationRequirementBoolean
-        configSpec =
-        new ConfigurationRequirementBoolean(expectedIdentifier);
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualNumberOfValues = configSpec.getMinNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedNumberOfValues, actualNumberOfValues);
-    assertTrue(configSpec.isFixNumberOfSettings());
-  }
-
-  /**
-   * Test method for {@link ConfigurationRequirementBoolean#ConfigurationRequirementBoolean(String,
-   * int)} <p/> The identifier should be set in the constructor and be retrievable through
-   * getIdentifier. The numberOfValues should be set to 2.
-   */
-  @Test
-  public void testConstructorGetTwo() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedNumberOfValues = 2;
-
-    // Execute functionality
-    ConfigurationRequirementBoolean
-        configSpec =
-        new ConfigurationRequirementBoolean(expectedIdentifier, expectedNumberOfValues);
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualNumberOfValues = configSpec.getMaxNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedNumberOfValues, actualNumberOfValues);
-    assertTrue(configSpec.isFixNumberOfSettings());
-  }
-
-
-  /**
-   * Test method for {@link ConfigurationRequirementBoolean#ConfigurationRequirementBoolean(String,
-   * int)} <p/> The identifier should be set in the constructor and be retrievable through
-   * getIdentifier. The numberOfValues should be set to the range (2, 4).
-   */
-  @Test
-  public void testConstructorGetRange() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedMinNumberOfValues = 2;
-    int expectedMaxNumberOfValues = 4;
-
-    // Execute functionality
-    ConfigurationRequirementBoolean
-        configSpec =
-        new ConfigurationRequirementBoolean(expectedIdentifier, expectedMinNumberOfValues, expectedMaxNumberOfValues);
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualMinNumberOfValues = configSpec.getMinNumberOfSettings();
-    int actualMaxNumberOfValues = configSpec.getMaxNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedMinNumberOfValues, actualMinNumberOfValues);
-    assertEquals(expectedMaxNumberOfValues, actualMaxNumberOfValues);
-    assertFalse(configSpec.isFixNumberOfSettings());
-  }
 
   /**
    * Test method for {@link ConfigurationRequirementBoolean#checkAndSetSettings(ConfigurationSettingBoolean...)}
@@ -140,7 +59,7 @@ public class ConfigurationRequirementBooleanTest {
   }
 
   /**
-   * Test method for {@link ConfigurationRequirementBoolean#checkAndSetDefaultValues(Boolean...)}
+   * Test method for {@link de.metanome.algorithm_integration.configuration.ConfigurationRequirementDefaultValue<Boolean>#checkAndSetDefaultValues(Boolean...)}
    *
    * The default values should be correctly settable on the specification.
    */
@@ -170,30 +89,6 @@ public class ConfigurationRequirementBooleanTest {
     // Check result
     assertEquals(true, configSpec.settings[0].value);
     assertEquals(false, configSpec.settings[1].value);
-  }
-
-  /**
-   * Test method for {@link ConfigurationRequirementBoolean#getDefaultValue(int)}
-   *
-   * The default values should be accessible via an index.
-   */
-  @Test
-  public void testGetDefaultValues() {
-    // Setup
-    ConfigurationRequirementBoolean
-        configSpec =
-        new ConfigurationRequirementBoolean("parameter1", 2);
-
-    // Execute functionality
-    try {
-      configSpec.checkAndSetDefaultValues(true, false);
-    } catch (AlgorithmConfigurationException e) {
-      fail();
-    }
-
-    // Check result
-    assertEquals(true, configSpec.getDefaultValue(0));
-    assertEquals(null, configSpec.getDefaultValue(2));
   }
 
   /**
