+++ algorithm_integration/src/test/java/de/metanome/algorithm_integration/configuration/ConfigurationRequirementFileInputTest.java
@@ -24,8 +24,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 
 /**
@@ -47,84 +45,6 @@ public class ConfigurationRequirementFileInputTest {
   @After
   public void tearDown() throws Exception {
 
-  }
-
-  /**
-   * Test method for {@link ConfigurationRequirementFileInput#ConfigurationRequirementFileInput(String)}
-   * <p/> The identifier should be set in the constructor and be retrievable through getIdentifier.
-   * The numberOfValues should be set to 1.
-   */
-  @Test
-  public void testConstructorGetOne() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedNumberOfValues = 1;
-    ConfigurationRequirementFileInput
-        configSpec =
-        new ConfigurationRequirementFileInput(expectedIdentifier);
-
-    // Execute functionality
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualNumberOfValues = configSpec.getMinNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedNumberOfValues, actualNumberOfValues);
-    assertTrue(configSpec.isFixNumberOfSettings());
-  }
-
-  /**
-   * Test method for {@link ConfigurationRequirementFileInput#ConfigurationRequirementFileInput(String,
-   * int)} <p/> The identifier should be set in the constructor and be retrievable through
-   * getIdentifier. The numberOfValues should be set to 2.
-   */
-  @Test
-  public void testConstructorGetTwo() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedNumberOFValues = 2;
-    ConfigurationRequirementFileInput
-        configSpec =
-        new ConfigurationRequirementFileInput(expectedIdentifier, expectedNumberOFValues);
-
-    // Execute functionality
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualNumberOfValues = configSpec.getMaxNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedNumberOFValues, actualNumberOfValues);
-    assertTrue(configSpec.isFixNumberOfSettings());
-  }
-
-  /**
-   * Test method for {@link ConfigurationRequirementFileInput#ConfigurationRequirementFileInput(String,
-   * int)} <p/> The identifier should be set in the constructor and be retrievable through
-   * getIdentifier. The numberOfValues should be set to the range (2, 4).
-   */
-  @Test
-  public void testConstructorGetRange() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedMinNumberOFValues = 2;
-    int expectedMaxNumberOFValues = 2;
-    ConfigurationRequirementFileInput
-        configSpec =
-        new ConfigurationRequirementFileInput(expectedIdentifier, expectedMinNumberOFValues, expectedMaxNumberOFValues);
-
-    // Execute functionality
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualMinNumberOfValues = configSpec.getMaxNumberOfSettings();
-    int actualMaxNumberOfValues = configSpec.getMaxNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedMaxNumberOFValues, actualMaxNumberOfValues);
-    assertEquals(expectedMinNumberOFValues, actualMinNumberOfValues);
-    assertTrue(configSpec.isFixNumberOfSettings());
   }
 
   /**
