+++ algorithm_integration/src/test/java/de/metanome/algorithm_integration/configuration/ConfigurationRequirementTableInputTest.java
@@ -24,9 +24,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 
 
@@ -46,84 +43,6 @@ public class ConfigurationRequirementTableInputTest {
   @After
   public void tearDown() throws Exception {
 
-  }
-
-  /**
-   * Test method for {@link ConfigurationRequirementTableInput#ConfigurationRequirementTableInput(String)}
-   * <p/> The identifier should be set in the constructor and be retrievable through getIdentifier.
-   * The numberOfValues should be set to 1.
-   */
-  @Test
-  public void testConstructorGetOne() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedNumberOfValues = 1;
-    ConfigurationRequirementTableInput
-        configSpec =
-        new ConfigurationRequirementTableInput(expectedIdentifier);
-
-    // Execute functionality
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualNumberOfValues = configSpec.getMinNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedNumberOfValues, actualNumberOfValues);
-    assertTrue(configSpec.isFixNumberOfSettings());
-  }
-
-  /**
-   * Test method for {@link ConfigurationRequirementTableInput#ConfigurationRequirementTableInput(String,
-   * int)} <p/> The identifier should be set in the constructor and be retrievable through
-   * getIdentifier. The numberOfValues should be set to 2.
-   */
-  @Test
-  public void testConstructorGetTwo() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedNumberOFValues = 2;
-    ConfigurationRequirementTableInput
-        configSpec =
-        new ConfigurationRequirementTableInput(expectedIdentifier, expectedNumberOFValues);
-
-    // Execute functionality
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualNumberOfValues = configSpec.getMinNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedNumberOFValues, actualNumberOfValues);
-    assertTrue(configSpec.isFixNumberOfSettings());
-  }
-
-  /**
-   * Test method for {@link ConfigurationRequirementTableInput#ConfigurationRequirementTableInput(String,
-   * int, int)} <p/> The identifier should be set in the constructor and be retrievable through
-   * getIdentifier. The numberOfValues should be set to the range (2, 4).
-   */
-  @Test
-  public void testConstructorGetRange() {
-    // Setup
-    // Expected values
-    String expectedIdentifier = "parameter1";
-    int expectedMinNumberOfValues = 2;
-    int expectedMaxNumberOfValues = 4;
-    ConfigurationRequirementTableInput
-        configSpec =
-        new ConfigurationRequirementTableInput(expectedIdentifier, expectedMinNumberOfValues, expectedMaxNumberOfValues);
-
-    // Execute functionality
-    String actualIdentifier = configSpec.getIdentifier();
-    int actualMinNumberOfValues = configSpec.getMinNumberOfSettings();
-    int actualMaxNumberOfValues = configSpec.getMaxNumberOfSettings();
-
-    // Check result
-    assertEquals(expectedIdentifier, actualIdentifier);
-    assertEquals(expectedMinNumberOfValues, actualMinNumberOfValues);
-    assertEquals(expectedMaxNumberOfValues, actualMaxNumberOfValues);
-    assertFalse(configSpec.isFixNumberOfSettings());
   }
 
   /**
